# Assembler version of Sum64String