--- a/xxhash_amd64.go
+++ b/xxhash_amd64.go
@@ -9,5 +9,11 @@ package xxhash
 //go:noescape
 func Sum64(b []byte) uint64
 
+// Sum64String computes the 64-bit xxHash digest of s.
+// It may be faster than Sum64([]byte(s)) by avoiding a copy.
+//
+//go:noescape
+func Sum64String(s string) uint64
+
 //go:noescape
 func writeBlocks(d *Digest, b []byte) int